AnalyticComb

Documentation for AnalyticComb.

AnalyticComb.W_coeff

AnalyticComb.p_binary_words_runl

AnalyticComb.W_coeff — Method

W_coeff(r;n_tot=200)

Taylor series coefficient from generating function for binary words that never have more than r consecutive identical letters.

The number of binary words that never have more than r consecutive identical letters is found to be (set α = β = r). n_tot defaults to 200, according to the example in Flajolet & Sedgewick pag. 52

AnalyticComb.p_binary_words_runl — Method

p_binary_word_runs(k,n)

Returns probablity associatied with k-lenght run in a sequence of size n.

Refer to the example in Flajolet & Sedgewick pag. 52.